refactor(turns): split the durable turn composition hub
`durable-turn-runtime.ts` owned five unrelated jobs at once: who is listening
to a turn, how a turn is generated, how a returning panel is caught up, what a
restart does with interrupted work, and which adapters the runtime is bound to.
Editing how a turn streams meant editing the module that decides who receives
it, and the shared in-memory maps meant nothing enforced the difference.

The pieces move to `src/background/turns/`:

- `turn-observers.ts` — delivery state and nothing else. Attach, buffer,
  forward, terminal cleanup, snapshots and reconnect leases. It imports no
  repository, provider, application or handler module, and a boundary test says
  so, because those are the dependencies that would pull generation back in.
- `turn-generation.ts` — provider invocation, stream reduction, assistant
  persistence.
- `turn-reconnect.ts` — snapshot assembly for a returning panel.
- `turn-recovery.ts` — stop intent, interrupted cancellations, resumption.
- `turn-service-factory.ts` — adapter binding, split out so recovery can build
  a service without importing the entry that re-exports it.

`durable-turn-runtime.ts` stays as the composition entry and keeps its import
path, so every caller is unchanged.

The synthetic port cast is gone with it. A stream producer needs `name`,
`postMessage`, and optionally an abort scope and a sequence counter — so that
is now `ChatStreamSink`, and the durable runtime declares one instead of
casting a three-property object to `ChromePort` and asserting an `onMessage`,
`onDisconnect`, `sender` and `disconnect()` that were never there. A real port
satisfies the sink structurally, `withErrorContext` is generic over the port
type with `ChromePort` as its default, and the one remaining cast adapts an
actual `browser.Runtime.Port` in `port-router.ts` — where a boundary test now
keeps it alone.

Behavior is unchanged: the existing durable-turn characterization suite covers
stream forwarding, reconnect snapshots, persistence and failure propagation,
and passes untouched.
