+++ AGENTS.md
@@ -286,6 +286,8 @@ Model-callable tools live in `src/lib/tools/internal/`, registered in `internal-
 
 `src/background/handlers/handle-{action}.ts`, registered in `src/background/index.ts`. Only streaming/port work belongs here (chat, context build, pull, selection actions, embedding download). Request/response provider and model operations live in `ProviderRpcService` / `ModelRpcService`. Keep handlers thin — adapt the port protocol to `src/lib/` and stream back.
 
+A handler that only *writes* a stream takes `ChatStreamSink`, not `ChromePort`. The sink is `name` + `postMessage` + the optional `abortScopeKey`/`streamSequence`, which is the whole surface a producer uses; a real port satisfies it structurally, so nothing at the port boundary changes. It exists because the durable turn runtime consumes the same stream in-process — it reduces events into durable state rather than shipping them to a panel — and used to reach `handleChatWithModel` by casting a three-property object to `ChromePort`, asserting an `onMessage`, `onDisconnect`, `sender` and `disconnect()` that did not exist. `withErrorContext` is generic over the port type and still defaults to `ChromePort`, so handlers that genuinely need a connection keep it. The one remaining `as unknown as ChromePort` is in `port-router.ts`, adapting a real `browser.Runtime.Port`, and a boundary test keeps it the only one.
+
 ### Component layers
 
 Four tiers, and the tier decides the rules:
@@ -458,5 +460,6 @@ What these files are *now*, so you neither go looking for a god-object that was 
 - `src/features/sessions/stores/chat-session-store.ts` is a ~19-LOC barrel over extracted slices; persistence reads via `chat-history.ts`. The old ~485-LOC store is gone.
 - `src/features/model/components/provider-settings.tsx` delegates connection details and custom model editing to small components. Keep new slices similarly scoped and covered by component tests.
 - `src/contents/index.ts` is a ~38-LOC entry; selection-capture, dom-observer, and messaging are siblings.
+- `src/background/durable-turn-runtime.ts` is a ~82-LOC composition entry — `startDurableTurn`, the live-only context callbacks, and re-exports that keep its import path stable. The pieces live in `src/background/turns/`: `turn-observers.ts` (delivery state — attach, buffer, forward, snapshots, reconnect leases, and nothing else), `turn-generation.ts` (provider invocation, stream reduction, assistant persistence), `turn-reconnect.ts` (snapshot assembly for a returning panel), `turn-recovery.ts` (stop intent, interrupted cancellations, restart resumption), `turn-service-factory.ts` (adapter binding, split out so recovery can build a service without importing the entry that re-exports it). `architecture-boundaries.test.ts` enforces the split: the registry imports no repository, provider, application or handler module, and its maps exist in exactly one file. Put a new control in the piece that owns it, not back in the entry.
 - `src/types/index.ts` is a re-export barrel (~11 LOC) over `chat`, `model`, `messaging`, `errors`, `content-extraction`, `ui-state`. Prefer the per-domain path (`@/types/chat`).
 - Dexie chat-history paths are retired. Vectors and knowledge sets still use Dexie; chat history is SQLite-only through the facade.
